src/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0    
aiohttp==3.7.4.post0
multidict==5.1.0
